Rename cache/test/ to cache/unit_tests/
